+++ src/App.tsx
@@ -9,9 +9,9 @@ import PBTIPage from "./pages/PBTIPage.tsx";
 import DiaryPage from "./pages/DiaryPage.tsx";
 import ProtectedRoute from "./components/common/ProtectedRoute.tsx";
 import KakaoSignupPage from "./pages/Login/KakaoSignupPage.tsx";
-import PerfumLabPage from "./pages/PerfumeLabPage.tsx";
-import LabLoadingPage from "./pages/LabLoadingPage.tsx";
-import LabResultPage from "./pages/LabResultPage.tsx";
+import PerfumLabPage from "./pages/perfumeLab/PerfumeLabPage.tsx";
+import LabLoadingPage from "./pages/perfumeLab/LabLoadingPage.tsx";
+import LabResultPage from "./pages/perfumeLab/LabResultPage.tsx";
 import ChatbotPage from "./pages/Chatbot/ChatbotPage.tsx";
 import ChoosePathPage from "./pages/ChoosePathPage.tsx";
 import ImageKeywordPage from "./pages/ImageKeyword/ImageKeywordPage.tsx";
@@ -34,8 +34,8 @@ const router = createBrowserRouter([
     element: <KakaoSignupPage />,
   },
   {
-  path: "/onboarding",
-  element: <OnboardingRouter  />
+    path: "/onboarding",
+    element: <OnboardingRouter />,
   },
   // 아래는 보호 라우트로 감싼 실제 서비스 페이지들
   {
@@ -47,23 +47,23 @@ const router = createBrowserRouter([
         errorElement: <NotFoundPage />,
         children: [
           { index: true, element: <MainPage /> },
-          { path: 'PBTI', element: <PBTIPage /> },
-          { path: 'Diary', element: <DiaryPage /> },
-          { path: 'MyPage', element: <MyPage /> },
-          { path: 'choose-path', element: <ChoosePathPage /> },
-          { path: 'filter', element: <FilterPage /> },
-          { 
-            path: 'image-keyword',
+          { path: "PBTI", element: <PBTIPage /> },
+          { path: "Diary", element: <DiaryPage /> },
+          { path: "MyPage", element: <MyPage /> },
+          { path: "choose-path", element: <ChoosePathPage /> },
+          { path: "filter", element: <FilterPage /> },
+          {
+            path: "image-keyword",
             children: [
               { index: true, element: <ImageKeywordPage /> },
-              { path: 'loading', element: <ImageKeywordLoading /> }
-            ]
-          }
-        ]
+              { path: "loading", element: <ImageKeywordLoading /> },
+            ],
+          },
+        ],
       },
       {
-        path: '/Chatbot',
-        element: <ChatbotPage />
+        path: "/Chatbot",
+        element: <ChatbotPage />,
       },
       {
         path: "/lab",
@@ -81,9 +81,9 @@ const router = createBrowserRouter([
             element: <LabResultPage />,
           },
         ],
-      }
-    ]
-  }
+      },
+    ],
+  },
 ]);
 
 function App() {
