+++ src/pages/perfumeLab/LabResultPage.tsx
